requirements.txt
anyio==4.4.0
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
fastapi==0.111.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
pandas==2.2.2
openpyxl==3.1.3